+++ changelog.md
@@ -1,4 +1,10 @@
 # version 1.0:
 
-- Added welcome and main pages;
+- Added "welcome" and "main" pages;
 - Added local storage save and load system;
+
+# version 1.1:
+
+- Imported JQuery.js library;
+- Added "edit link" page;
+- Added edit links system;